==> experiments/04-Apply-med7-on-Clinical-Notes_Updated.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

DATAPATH = "/content/drive/MyDrive/Colab Notebooks/ConvolutionMedicalNer-master/data/"

import pandas as pd
import os
#!pip install https://huggingface.co/kormilitzin/en_core_med7_lg.tar.gz
!pip install https://huggingface.co/kormilitzin/en_core_med7_lg/resolve/main/en_core_med7_lg-any-py3-none-any.whl
!pip install -U spacy==3.2.0
#import spacy.cli
#spacy.cli.download("en_core_med7_lg")
#spacy.cli.download("en_core_med7_trf")
import spacy
#import en_core_med7_lg
med7 = spacy.load('en_core_med7_lg')

#preprocessed_df = pd.read_pickle("/content/drive/MyDrive/Colab Notebooks/ConvolutionMedicalNer-master/data/preprocessed_notes.p")
preprocessed_df = pd.read_pickle(os.path.join(DATAPATH, "preprocessed_notes_ablation.p"))

preprocessed_df['ner'] = None

print(len(preprocessed_df))

count = 0
preprocessed_index = {}
for i in preprocessed_df.itertuples():
    
    if count % 1000 == 0:
        print(count)
        #if count == 1000:
        #    break  

    count += 1
    ind = i.Index
    text = i.preprocessed_text
    
    all_pred = []
    for each_sent in text:
        try:
            doc = med7(each_sent)
            result = ([(ent.text, ent.label_) for ent in doc.ents])
            if len(result) == 0: continue
            all_pred.append(result)
        except:
            print("error..")
            continue
    preprocessed_df.at[ind, 'ner'] = all_pred

#pd.to_pickle(preprocessed_df, "/content/drive/MyDrive/Colab Notebooks/ConvolutionMedicalNer-master/data/ner_df_final.p")
pd.to_pickle(preprocessed_df, (os.path.join(DATAPATH, "ner_df_final_ablation.p")))
print("Completed Writing Final Ablation File")
